kuber/services/Inference/src/requirements.txt
requests==2.32.3
onnx==1.16.0
onnxruntime==1.17.1
aiofiles
onnx2torch
